requirements.txt
Django>=1.8.2
django-configurations>=0.8
pyYAML>=3.10
suit>=2.0.2
pytz>=2015.2
pyzmq>=14.5.0
djangorestframework>=3.1.3